fix(gcs): Conditionally load GCS credentials
This commit fixes an issue where the application would fail to start in a Cloud Run environment because it was trying to load GCS credentials from an environment variable that was not present. The `config/storage.yml` has been updated to only load the credentials if the `GCS_CREDENTIALS_JSON` environment variable is present.

Fixes #15